+++ NEWS.md
@@ -1,5 +1,7 @@
 # ussie 0.0.0.9000
 
-* Changed `_games` to `_matches`, e.g. `uss_make_matches()`.
+* Added `uss_make_teams_matches()` to convert a `matches` tibble to a `teams-matches` tibble. (#1)
+
+* Changed `_games` to `_matches`, e.g. `uss_make_matches()`. (#2)
 
 * Added a `NEWS.md` file to track changes to the package.